requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.14.0
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docutils==0.21.2
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
libsass==0.23.0
libsql-client==0.3.1
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sqlalchemy-libsql==0.1.0
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.2
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.18.3
